refactor(solver): split db base cache traits
