+++ crates/tsz-solver/src/caches/mod.rs
@@ -1,5 +1,6 @@
 mod application_eval_index;
 pub(crate) mod db;
+mod db_base_traits;
 pub(crate) mod display_provenance;
 pub(crate) mod instantiation_cache;
 pub(crate) mod options;
